ReadBlog Page & Display Blog Page:  Code Refactor && code cleaning
