+++ lib/frontend/user/pages/blogs/displayblogs/widgets/check_internet.dart/no_internet.dart
