+++ lib/main.dart
@@ -5,7 +5,7 @@ import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:rafay_portfolio/backend/auth/auth_gate.dart';
 import 'package:rafay_portfolio/firebase_options.dart';
-import 'package:rafay_portfolio/frontend/user/pages/blogs/read_blog.dart';
+import 'package:rafay_portfolio/frontend/user/pages/blogs/readblogs/read_blog.dart';
 import 'package:rafay_portfolio/frontend/user/pages/error/page404.dart';
 import 'package:url_strategy/url_strategy.dart';
 import 'package:rafay_portfolio/constants/routes/routes.dart';
@@ -16,12 +16,10 @@ import 'package:firebase_analytics/firebase_analytics.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  // Firebase init with platform
   await Firebase.initializeApp(
     options: DefaultFirebaseOptions.currentPlatform,
   );
-
-  // firebase Error Catching and Crash Reporting
+  await FirebaseAnalytics.instance.setAnalyticsCollectionEnabled(true);
   FlutterError.onError = FirebaseCrashlytics.instance.recordFlutterFatalError;
   WidgetsFlutterBinding.ensureInitialized();
 
@@ -32,21 +30,15 @@ Future<void> main() async {
     FirebaseCrashlytics.instance.recordError(error, stack, fatal: true);
     return true;
   };
-
-  //setup URLs
   setPathUrlStrategy();
-
-  // Running the application
   runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
-  // Firebase Analytics
   static FirebaseAnalytics analytics = FirebaseAnalytics.instance;
   static FirebaseAnalyticsObserver observer =
       FirebaseAnalyticsObserver(analytics: analytics);
-  //Main Widget
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -69,28 +61,11 @@ class MyApp extends StatelessWidget {
             ),
           );
         }
-        // Unknown route
         return MaterialPageRoute(builder: (context) => const Page404());
       },
       onUnknownRoute: (settings) => MaterialPageRoute(
         builder: (context) => const Page404(),
       ),
     );
-
-    // Orginal Code without onGenerateRoute & Auto Routes
-    // return MaterialApp(
-    //   navigatorObservers: <NavigatorObserver>[observer],
-    //   debugShowCheckedModeBanner: false,
-    //   theme: lightMode,
-    //   initialRoute: '/',
-    //   routes: {
-    //     '/': (context) => const SplashScreen(),
-    //     '/admin': (context) => const AuthGate(),
-    //     ...appRoutes,
-    //   },
-    //   onUnknownRoute: (settings) => MaterialPageRoute(
-    //     builder: (context) => const Page404(),
-    //   ),
-    // );
   }
 }
