+++ pubspec.yaml
@@ -49,6 +49,8 @@ dependencies:
   dio: ^5.4.2
   html_parser_plus: ^0.1.4
   highlight: ^0.7.0
+  share: ^2.0.4
+  share_plus: ^8.0.2
 dev_dependencies:
   flutter_test:
     sdk: flutter
