+++ web/index.html
@@ -18,7 +18,10 @@
     <base href="/" />
     <meta charset="UTF-8" />
     <meta content="IE=Edge" http-equiv="X-UA-Compatible" />
-    <meta name="description" content="A new Flutter project." />
+    <meta
+      name="description"
+      content="Discover the dynamic world of Abdul Rafay—a fusion of creativity, innovation, and passion. Explore captivating blog posts and stunning portfolio showcases. Join me on a journey of inspiration, collaboration, and endless possibilities. Let's shape the future, one idea at a time."
+    />
 
     <!-- iOS meta tags & icons -->
     <meta name="apple-mobile-web-app-capable" content="yes" />
@@ -32,7 +35,7 @@
     <title>Abdul Rafay</title>
     <link rel="manifest" href="manifest.json" />
 
-    <script>
+    <script defer>
       // The value below is injected by flutter build, do not touch.
       const serviceWorkerVersion = null;
     </script>
@@ -40,7 +43,7 @@
     <script src="flutter.js" defer></script>
   </head>
   <body>
-    <script>
+    <script defer>
       window.addEventListener("load", function (ev) {
         // Download main.dart.js
         _flutter.loader.loadEntrypoint({
